data_generation/requirements.txt
openai>=1.0.0
PyYAML>=6.0
json-repair>=0.15.0
pandas>=1.5.0 